Rearrange: move code into numbers feature module
